SignalExtOption
